Add integration tests
