+++ tests/BookMe.IntegrationTests/BookingTests.cs
@@ -45,147 +45,6 @@ public class BookingTests : BaseIntegrationTest
         _bookingController = new BookingController(_mediator, _timeSlotQueries);
     }
 
-    #region CreateTimeSlot tests
-    [Fact]
-    public async Task CreateTimeSlotsShouldSucceedAsync()
-    {
-        // Arrange
-        await _bookMeContext.TimeSlots.ExecuteDeleteAsync();
-        _mockHttpContext.SetUser(_adminUser);
-
-        var createTimeSlotsRequest = new CreateTimeSlotsDto
-        {
-            StartDateTime = DateTime.UtcNow.AddDays(10).AddHours(1),
-            EndDateTime = DateTime.UtcNow.AddDays(10).AddHours(2),
-            IsAllDay = false,
-        };
-
-        // Act
-        var result = await _bookingController.CreateTimeSlotsAsync(createTimeSlotsRequest);
-
-        // Assert
-        result.ValidateOkResult<List<TimeSlotDto>>(timeSlots =>
-        {
-            var timeSlot = timeSlots.First();
-            timeSlot.Start.Should().Be(createTimeSlotsRequest.StartDateTime);
-            timeSlot.End.Should().Be(createTimeSlotsRequest.EndDateTime);
-            timeSlot.IsAvailable.Should().BeTrue();
-            timeSlot.Id.Should().NotBeEmpty();
-        });
-
-        var timeSlots = await _bookMeContext.TimeSlots.ToListAsync();
-
-        timeSlots.Should().HaveCount(1);
-    }
-
-    [Fact]
-    public async Task CreateOverlappingTimeSlotsShouldNotSucceedAsync()
-    {
-        // Arrange
-        await _bookMeContext.TimeSlots.ExecuteDeleteAsync();
-        _mockHttpContext.SetUser(_adminUser);
-
-        var firstCreateTimeSlotsRequest = new CreateTimeSlotsDto
-        {
-            StartDateTime = DateTime.UtcNow.AddDays(10).AddHours(1),
-            EndDateTime = DateTime.UtcNow.AddDays(10).AddHours(2),
-            IsAllDay = false,
-        };
-
-        var firstResult = await _bookingController.CreateTimeSlotsAsync(
-            firstCreateTimeSlotsRequest
-        );
-
-        var createTimeSlotsRequest = new CreateTimeSlotsDto
-        {
-            StartDateTime = DateTime.UtcNow.AddDays(10).AddHours(1),
-            EndDateTime = DateTime.UtcNow.AddDays(10).AddHours(2),
-            IsAllDay = false,
-        };
-
-        // Act
-        var result = await _bookingController.CreateTimeSlotsAsync(createTimeSlotsRequest);
-
-        // Assert
-        result.ValidateBadRequestResult<List<Error>>(errors =>
-        {
-            errors.Should().NotBeEmpty();
-            var firstError = errors.First();
-            firstError.Should().NotBeNull();
-            firstError.Code.Should().Be("conflict");
-            firstError.Description
-                .Should()
-                .Be("The requested time slot overlaps with existing time slots");
-        });
-
-        var timeSlots = await _bookMeContext.TimeSlots.ToListAsync();
-
-        timeSlots.Should().HaveCount(1);
-    }
-
-    [Fact]
-    public async Task CreateTimeSlotsWithNonAdminUser_ShouldNotSucceedAsync()
-    {
-        // Arrange
-        await _bookMeContext.TimeSlots.ExecuteDeleteAsync();
-        var customerUser = await _bookMeContext.Users
-            .Include(x => x.UserRoles)
-            .ThenInclude(x => x.Role)
-            .FirstAsync(x => x.UserRoles.Any(y => y.RoleId == DefaultRoles.CustomerId));
-
-        _mockHttpContext.SetUser(customerUser.MapToDto());
-
-        var createTimeSlotsRequest = new CreateTimeSlotsDto
-        {
-            StartDateTime = DateTime.UtcNow.AddDays(10).AddHours(1),
-            EndDateTime = DateTime.UtcNow.AddDays(10).AddHours(2),
-            IsAllDay = false,
-        };
-
-        // Act & Assert
-        var exception = await Assert.ThrowsAsync<ValidationException>(
-            () => _bookingController.CreateTimeSlotsAsync(createTimeSlotsRequest)
-        );
-
-        exception.Message
-            .Should()
-            .Be($"Validation failed: \n -- UserDto: User {customerUser.Id} is not an admin");
-
-        var timeSlots = await _bookMeContext.TimeSlots.ToListAsync();
-
-        timeSlots.Should().HaveCount(0);
-    }
-
-    [Theory]
-    [InlineData(5)]
-    [InlineData(10)]
-    [InlineData(30)]
-    public async Task CreateTimeSlotWithDifferentDays_ShouldSucceedAsync(int daysInFuture)
-    {
-        // Arrange
-        await _bookMeContext.TimeSlots.ExecuteDeleteAsync();
-        _mockHttpContext.SetUser(_adminUser);
-
-        var createTimeSlotsRequest = new CreateTimeSlotsDto
-        {
-            StartDateTime = DateTime.UtcNow.AddDays(daysInFuture).AddHours(1),
-            EndDateTime = DateTime.UtcNow.AddDays(daysInFuture).AddHours(2),
-        };
-
-        // Act
-        var result = await _bookingController.CreateTimeSlotsAsync(createTimeSlotsRequest);
-
-        // Assert
-        result.ValidateOkResult<List<TimeSlotDto>>(timeSlots =>
-        {
-            timeSlots.Should().HaveCount(1);
-            var timeSlot = timeSlots.First();
-            timeSlot.Start.Should().Be(createTimeSlotsRequest.StartDateTime);
-            timeSlot.End.Should().Be(createTimeSlotsRequest.EndDateTime);
-        });
-    }
-    #endregion
-
     #region BookTimeSlot tests
     [Fact]
     public async Task BookTimeSlotShouldWithCustomerUser_ShouldSucceedAsync()
